+++ tests/ImboUnitTest/EventListener/StatsAccessTest.php
@@ -10,7 +10,10 @@
 
 namespace ImboUnitTest\EventListener;
 
-use Imbo\EventListener\StatsAccess;
+use Imbo\EventListener\StatsAccess,
+    Imbo\Resource\Stats as StatsResource,
+    Imbo\EventManager\EventManager,
+    ReflectionProperty;
 
 /**
  * @covers Imbo\EventListener\StatsAccess
@@ -151,5 +154,44 @@ class StatsAccessTest extends ListenerTests {
         }
 
         $listener->checkAccess($this->event);
+    }
+
+    /**
+     * @see https://github.com/imbo/imbo/issues/249
+     */
+    public function testListensToTheSameEventsAsTheStatsResource() {
+        $this->assertSame(
+            array_keys(StatsAccess::getSubscribedEvents()),
+            array_keys(StatsResource::getSubscribedEvents()),
+            'The stats access event listener does not listen to the same events as the stats resource, which it should'
+        );
+    }
+
+    /**
+     * @see https://github.com/imbo/imbo/issues/251
+     *
+     * This test is best run with "In the Ghetto" blasting on the stereo
+     */
+    public function testHasHigherPriorityThanTheStatsResource() {
+        $statsAccess = new StatsAccess();
+        $statsResource = new StatsResource();
+
+        $eventManager = new EventManager();
+        $eventManager->addEventHandler('statsAccess', $statsAccess);
+        $eventManager->addCallbacks('statsAccess', StatsAccess::getSubscribedEvents());
+        $eventManager->addEventHandler('statsResource', $statsResource);
+        $eventManager->addCallbacks('statsResource', StatsResource::getSubscribedEvents());
+
+        $callbacks = new ReflectionProperty($eventManager, 'callbacks');
+        $callbacks->setAccessible(true);
+
+        $handlersForGet = $callbacks->getValue($eventManager)['stats.get'];
+        $handlersForHead = $callbacks->getValue($eventManager)['stats.head'];
+
+        $this->assertSame($statsAccess, $eventManager->getHandlerInstance($handlersForGet->extract()['handler']));
+        $this->assertSame($statsResource, $eventManager->getHandlerInstance($handlersForGet->extract()['handler']));
+
+        $this->assertSame($statsAccess, $eventManager->getHandlerInstance($handlersForHead->extract()['handler']));
+        $this->assertSame($statsResource, $eventManager->getHandlerInstance($handlersForHead->extract()['handler']));
     }
 }